+++ core/CHANGELOG.md
@@ -9,10 +9,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - Deref `ItemCollection` ([#363](https://github.com/stac-utils/stac-rs/pull/363))
+- `object_store` and `ndjson` ([#369](https://github.com/stac-utils/stac-rs/pull/369))
 
 ### Changed
 
 - Update **geoarrow** to v0.3.0 ([#367](https://github.com/stac-utils/stac-rs/pull/367))
+
+### Removed
+
+- `io` module ([#369](https://github.com/stac-utils/stac-rs/pull/369))
 
 ## [0.9.0] - 2024-09-05